refactor: 체스판 생성 로직 분리
