+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import chess.domain.chessboard.Chessboard;
+import chess.domain.chessboard.ChessboardFactory;
 import chess.domain.piece.attribute.Color;
 import chess.domain.piece.attribute.Position;
 
@@ -16,7 +16,7 @@ class PawnTest {
     @Test
     void move() {
         Piece sut = new Pawn(Color.WHITE, Position.from("d3"));
-        Position actual = sut.move(Chessboard.empty(), Position.from("d4")).position();
+        Position actual = sut.move(ChessboardFactory.empty(), Position.from("d4")).position();
         assertThat(actual).isEqualTo(Position.from("d4"));
     }
 
@@ -24,7 +24,7 @@ class PawnTest {
     @Test
     void moveException() {
         Piece sut = new Pawn(Color.WHITE, Position.from("d5"));
-        assertThatCode(() -> sut.move(Chessboard.empty(), Position.from("e3")))
+        assertThatCode(() -> sut.move(ChessboardFactory.empty(), Position.from("e3")))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("이동할 수 없는 위치입니다: E3");
     }
