+++ src/test/java/chess/domain/piece/attribute/MovementTest.java
@@ -6,11 +6,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class MovementTest {
-	@DisplayName("방향을 1개 이상 입력하지 않으면 예외가 발생한다.")
-	@Test
-	void of() {
-		assertThatThrownBy(() -> Movement.of())
-				.isInstanceOf(IllegalArgumentException.class)
-				.hasMessage("이동을 위한 방향은 1개 이상입니다.");
-	}
+    @DisplayName("방향을 1개 이상 입력하지 않으면 예외가 발생한다.")
+    @Test
+    void of() {
+        assertThatThrownBy(Movement::of)
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("이동을 위한 방향은 1개 이상입니다.");
+    }
 }
